【杂项】见详细
1.WndRanking调整装备栏格子大小
2.CharSprite同一状态特效只允许存在一个
3.辅助鉴定加入“展示原因”
4.局内放大镜查看物品时，物品以堆形式存在，并且该格有多个物品，则以物品窗查看。
5.炼金产生鉴定时展示鉴定名字。
6.井房的水晶钥匙放在最上面。
7. .50加入unique属性，以防被小偷偷掉或者被炸掉
8.GM6及SRS、AN94的成长上调0.5，AN94对索敌附魔也双判。
9.M99攻速调整到1回合。
10.稍微调整Cypros的数值。
11.液金允许修改defaultAction以便在快捷栏进行丢出
12.蜕变时将其中1点天赋点转换回自由天赋点（没有加点则不转换）

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/scenes/PixelScene.java
@@ -256,12 +256,12 @@ public class PixelScene extends Scene {
 	}
 
 
-	public static RenderedTextBlock renderTextBlock(int size ){
+	public static RenderedTextBlock renderTextBlock(float size ){
 		return renderTextBlock("", size);
 	}
 
-	public static RenderedTextBlock renderTextBlock(String text, int size ){
-		RenderedTextBlock result = new RenderedTextBlock( text, size*defaultZoom);
+	public static RenderedTextBlock renderTextBlock(String text, float size ){
+		RenderedTextBlock result = new RenderedTextBlock( text, (int) (size*defaultZoom));
 		result.zoom(1/(float)defaultZoom);
 		return result;
 	}
